Update the README for devops interviews.
Provide information on what the candidate's laptop should be configured to do.
The text is deliberately written not to be prescriptive so candidates have the
freedom to use whatever tooling they prefer as long as they can work on the
exercise.

Remove explicit command snippets since most candidates are fluent with the
commands.

Provide a more fleshed out list of proposed tools/util which are commonly in use
today.